disable snapshots
Currently there is a limitation in Azure SDK and snapshots can only
be created, not restored. We don't want users to be able to create
snapshots when there is no use for them at this point.

Disable snapshots by removing volume snapshot class resource and
removing snapshotter sidecars.

PR for tracking snapshot restore:
https://github.com/kubernetes-sigs/azurefile-csi-driver/issues/136