simplify workload identity feature enablement
Injector is now changed so it's safe to set --enable-azure-workload-identity
to "true" even when some of the values are not present. In that case
injector will try to revert to user+secret authentication and log a
warning.

Setting --enable-azure-workload-identity to "false" will make injector
always use user+secret ignoring any values for workload identity that
might be present.